move password collecting to config cache
